[util,basegen] Move class Name to basegen
The Name class has no dependencies and is pretty useful, so it is
best to place it under basegen.
Add BUILD file in basegen.
