+++ util/dtgen/dt_api.c.tpl
@@ -4,7 +4,8 @@
 //
 // Device table API auto-generated by `dtgen`
 <%
-from topgen.lib import Name, is_top_reggen, is_ipgen
+from basegen.lib import Name
+from topgen.lib import is_top_reggen, is_ipgen
 %>
 
 /**
