+++ util/dtgen/dt_ip.h.tpl
@@ -4,7 +4,7 @@
 //
 // Device table API auto-generated by `dtgen`
 <%
-    from topgen.lib import Name
+    from basegen.lib import Name
     from dtgen.helper import Extension
 
     device_name = helper.ip.name
